Split kanji search by radical functions.
Put them in another file in preparation for adding sentence searching
which will add considerable complexity to the code.
